Dev/mid

+++ server/src/main/java/com/asap/server/dto/response/CodeBattleMatchResult.java
@@ -11,5 +11,5 @@ public class CodeBattleMatchResult {
     private Long matchId;
     private int winner;
     private String log;
-    private Integer aiOrder;
+    private Long aiOrder;
 }
